Move obs mocks to observe lib
