+++ observelib/observe/src/main/ballerina/Ballerina.toml
@@ -2,6 +2,7 @@
 org = "ballerina"
 name = "observe"
 version = "0.9.0"
+export = ["observe", "observe.mockextension" ]
 
 [[platform.java11.dependency]]
 path = "../libs/ballerina-observability-2.0.0-alpha7-SNAPSHOT.jar"
