+++ observelib/observe/src/main/ballerina/modules/mockextension/tracing.bal
@@ -31,5 +31,5 @@ public isolated function getFinishedSpans(string serviceName) returns json {
 # + return - The finished spans as a json
 isolated function externGetFinishedSpans(handle serviceName) returns json = @java:Method {
     name: "getFinishedSpans",
-    'class: "org.ballerina.testobserve.tracing.extension.MockTracerUtils"
+    'class: "org.ballerinalang.observe.mockextension.MockTracerUtils"
 } external;
