+++ observelib/observe/src/main/java/org/ballerinalang/observe/mockextension/BMockTracerProvider.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerina.testobserve.tracing.extension;
+package org.ballerinalang.observe.mockextension;
 
 import io.ballerina.runtime.observability.tracer.spi.TracerProvider;
 import io.opentelemetry.api.trace.Tracer;
