+++ observelib/observe/src/main/resources/META-INF/services/io.ballerina.runtime.observability.tracer.spi.TracerProvider
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.ballerina.testobserve.tracing.extension.BMockTracerProvider
+org.ballerinalang.observe.mockextension.BMockTracerProvider
