+++ tests/jballerina-integration-test/src/test/java/org/ballerinalang/test/observability/metrics/MetricsTestCase.java
@@ -23,9 +23,9 @@ import com.google.gson.reflect.TypeToken;
 import io.ballerina.runtime.observability.metrics.PercentileValue;
 import io.ballerina.runtime.observability.metrics.Snapshot;
 import io.ballerina.runtime.observability.metrics.Tag;
-import org.ballerina.testobserve.metrics.extension.model.Metrics;
-import org.ballerina.testobserve.metrics.extension.model.MockGauge;
-import org.ballerina.testobserve.metrics.extension.model.MockMetric;
+import org.ballerinalang.observe.mockextension.model.Metrics;
+import org.ballerinalang.observe.mockextension.model.MockGauge;
+import org.ballerinalang.observe.mockextension.model.MockMetric;
 import org.ballerinalang.test.observability.ObservabilityBaseTest;
 import org.ballerinalang.test.util.HttpClientRequest;
 import org.ballerinalang.test.util.HttpResponse;
