--- tests/jballerina-integration-test/src/test/resources/configurables/configLibProject/Dependencies.toml
@@ -1,1 +1,0 @@
-
