--- tests/jballerina-integration-test/src/test/resources/configurables/invalidDefaultable/main/Dependencies.toml
@@ -1,1 +1,0 @@
-
