+++ tests/jballerina-integration-test/src/test/resources/observability/metrics_tests/commons.bal
@@ -14,11 +14,12 @@
 // specific language governing permissions and limitations
 // under the License.
 
+import ballerina/observe.mockextension;
 import ballerina/testobserve;
 
 service /metricsRegistry on new testobserve:Listener(10090) {
     resource function post getMetrics(testobserve:Caller caller) {
-        json metrics = testobserve:getMetrics();
+        json metrics = mockextension:getMetrics();
         checkpanic caller->respond(metrics.toJsonString());
     }
 }
