+++ tests/jballerina-integration-test/src/test/resources/observability/tracing_tests/commons.bal
@@ -15,6 +15,7 @@
 // under the License.
 
 import ballerina/observe;
+import ballerina/observe.mockextension;
 import ballerina/testobserve;
 import intg_tests/tracing_tests.utils as utils;
 
@@ -23,7 +24,7 @@ utils:MockClient testClient = new();
 @display { label: "mockTracer" }
 service /mockTracer on new testobserve:Listener(19090) {
     resource function post getMockTraces(testobserve:Caller caller, string serviceName) {
-        json spans = testobserve:getFinishedSpans(serviceName);
+        json spans = mockextension:getFinishedSpans(serviceName);
         checkpanic caller->respond(spans.toJsonString());
     }
 }
